restructured ARU classes into their own package
