+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/aru/AruProduct.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2020, Oracle Corporation and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.weblogic.imagetool.installer;
+package com.oracle.weblogic.imagetool.aru;
 
 /**
  * ARU product codes for patching.
